# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "description": "Error is a short, human-readable description of what went\nwrong. The server never returns raw stack traces or DB\nerror strings here; the goal is to be informative without\nleaking internals.",
                        "example": "Invalid request payload",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "comment.CommentListResponse": {
                "properties": {
                    "comments": {
                        "items": {
                            "$ref": "#/components/schemas/model.Comment"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "comment.CommentModerationListResponse": {
                "properties": {
                    "comments": {
                        "items": {
                            "$ref": "#/components/schemas/model.Comment"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/comment.Pagination"
                    }
                },
                "type": "object"
            },
            "comment.CreateCommentRequest": {
                "properties": {
                    "content": {
                        "example": "Great post!",
                        "maxLength": 100,
                        "type": "string"
                    },
                    "parentId": {
                        "example": 7,
                        "type": "integer"
                    },
                    "postId": {
                        "example": 42,
                        "type": "integer"
                    }
                },
                "required": [
                    "content",
                    "postId"
                ],
                "type": "object"
            },
            "comment.CreateCommentResponse": {
                "properties": {
                    "comment": {
                        "$ref": "#/components/schemas/model.Comment"
                    },
                    "commentsCount": {
                        "example": 3,
                        "type": "integer"
                    },
                    "message": {
                        "example": "Comment created successfully",
                        "type": "string"
                    },
                    "requiresModeration": {
                        "example": false,
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "comment.DeleteCommentResponse": {
                "properties": {
                    "commentsCount": {
                        "example": 3,
                        "type": "integer"
                    },
                    "message": {
                        "example": "Comment deleted",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "JWT bearer token issued by /api/auth/login",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "contact": {
            "name": "GitHub Issues",
            "url": "https://github.com/vexgo-org/vexgo/issues"
        },
        "description": "Self-hosted blog CMS HTTP API.",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "termsOfService": "https://github.com/vexgo-org/vexgo",
        "title": "VexGo API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/comments": {
            "post": {
                "description": "Adds a comment to a post. Content is capped at 100\ncharacters. The postId accepts either a number or a\nstring. The reply is held for moderation when the\nmanual review queue is on, the keyword filter rejects\nit, or the LLM filter rejects it.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/comment.CreateCommentRequest",
                                        "summary": "request",
                                        "description": "comment payload"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "comment payload",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/comment.CreateCommentResponse"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "validation error or invalid postId"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "not logged in"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create a comment",
                "tags": [
                    "comments"
                ]
            }
        },
        "/comments/post/{id}": {
            "get": {
                "description": "Returns all published comments for the post. Pending\ncomments are only visible to the author and to admins\n(the service layer filters by role).",
                "parameters": [
                    {
                        "description": "post id or slug",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/comment.CommentListResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "List comments for a post",
                "tags": [
                    "comments"
                ]
            }
        },
        "/comments/{id}": {
            "delete": {
                "description": "The author or an admin may delete a comment.",
                "parameters": [
                    {
                        "description": "comment id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/comment.DeleteCommentResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "not logged in"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "not author or admin"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "comment not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete a comment",
                "tags": [
                    "comments"
                ]
            }
        },
...
        "/moderation/comments/approved": {
            "get": {
                "parameters": [
                    {
                        "description": "page number (1-based)",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 1,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "page size",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "default": 10,
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/comment.CommentModerationListResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List approved comments",
                "tags": [
                    "comments"
                ]
            }
        },
        "/moderation/comments/config": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model.CommentModerationConfig"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get comment moderation config",
                "tags": [
                    "comments"
                ]
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/comment.UpdateModerationConfigBody",
                                        "summary": "request",
                                        "description": "new config"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "new config",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/comment.UpdateModerationConfigResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "incomplete LLM config"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Update comment moderation config",
                "tags": [
                    "comments"
                ]
            }
        },
        "/moderation/comments/pending": {
            "get": {
                "parameters": [
                    {
                        "description": "page number (1-based)",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 1,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "page size",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "default": 10,
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/comment.CommentModerationListResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List pending comments (moderation queue)",
                "tags": [
                    "comments"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
...
}